libhurd-ihash/
2008-01-23  Neal H. Walfield  <neal@gnu.org>

	* ihash.h (hurd_ihash_buffer_size): New declaration.
	(hurd_ihash_init_with_buffer): Likewise.
	(hurd_ihash_init): Add comment that this function is not provided
	if the library is compiled with NO_MALLOC.
	(hurd_ihash_create): Likewise.
	(hurd_ihash_free): Likewise.
	(hurd_ihash_destroy): Update comment.
	* ihash.c (hurd_ihash_init_with_buffer): New function.
	(hurd_ihash_destroy) [NO_MALLOC]: Don't free HT->ITEMS.
	(hurd_ihash_create) [NO_MALLOC]: Just return ENOMEM;
	(hurd_ihash_free) [NO_MALLOC]: Remove function.
	(hurd_ihash_buffer_size): New function.
	(hurd_ihash_replace) [NO_MALLOC]: Don't attempt to expand the hash
	table.
	[! NO_MALLOC]: Use hurd_ihash_buffer_size to calculate an
	appropriate buffer size.
	* Makefile.am (lib_LIBRARIES): Add libhurd-ihash-nomalloc.a.
	(libhurd_ihash_nomalloc_a_CPPFLAGS): New variable.
	(libhurd_ihash_nomalloc_a_SOURCES): Likewise.

	* t-ihash.c (main): Cast locp offset calculation to elide warning.

viengoos/

2008-01-23  Neal H. Walfield  <neal@gnu.org>

	* thread.c: Include "zalloc.h".
	(thread_init): Initialize TID_TO_THREAD hash here.  Allocate using
	zalloc.
	* object.c: Include "zalloc.h".
	(object_init): Allocate hash table and object_descs using zalloc.
	* Makefile.am (viengoos_SOURCES): Remove malloc.h and
	malloc-wrap.c.
	(viengoos_LDADD): Use ../libhurd-ihash/libhurd-ihash-nomalloc.a
	rather than ../libhurd-ihash/libhurd-ihash.a.
	(EXTRA_viengoos_SOURCES): Remove variable.

	* memory.c (memory_grab) [_L4_TEST_ENVIRONMENT]: Don't include
	binary when calculating available memory.

+++ libhurd-ihash/t-ihash.c
@@ -104,7 +104,7 @@ main (int argc, char *argv[])
     int value;
     hurd_ihash_locp_t locp;
   };
-  hurd_ihash_init (&hash, &(((struct s *)0)->locp));
+  hurd_ihash_init (&hash, (int) &(((struct s *)0)->locp));
 
   if (hurd_ihash_find (&hash, 1))
     F ("Found object with key 1 in otherwise empty hash");
